Add/Delete/Edit product are all now working.
I think it's a little hacky, but it'll do for now.  If someone wants to clean up behind me if we have time awesome, and hopefully it'll still integrate well with everything else that needs to be done.

All that's left for project 1 is the validation of input.  This can be done with JS if someone could take care of this, but this now unblocks projects 2+